tray_mcp_server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0
alembic>=1.13.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
redis>=5.0.1
celery>=5.3.4
aiofiles>=23.2.1
jinja2>=3.1.2
python-dateutil>=2.8.2
requests>=2.31.0
